LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/index-not-in-entries-expected.txt



FAIL entry.index should return -1 when not in navigation.entries() assert_equals: expected -1 but got 1
